+++ runme.py
@@ -10,7 +10,7 @@ from expworkup import jsontocsv
 from expworkup import createjson
 from expworkup import googleio
 from versiondata import export_to_repo
-from expworkup.handlers import compound_ingredients
+from expworkup.entity_tables import reagent_entity
 from tests import logger
 from utils import globals
 
@@ -53,8 +53,8 @@ if __name__ == "__main__":
                         help='generates <target folder>_models.csv and <target_folder>_objects.csv\
                         corresponding to all compound ingredients in the target dataset ||default = False||')
     parser.add_argument('-v', '--verdata', type=str, 
-                        help='generates <XXXX>.perovskitedata.csv output in a form ready for upload to the \
-                        versioned data repo ||default = None||')
+                        help='Enter numerical value such as "0001". Generates <0001>.perovskitedata.csv output\
+                        in a form ready for upload to the versioned data repo ||default = None||')
     parser.add_argument('-s', '--state', type=str,
                         help='title of state set file to be used at the state set for \
                         this iteration of the challenge problem, no entry will result in no processing')
@@ -83,19 +83,20 @@ if __name__ == "__main__":
 
     initialize(args)
 
-    chem_df=googleio.ChemicalData()  #Grabs relevant chemical data frame from google sheets (only once no matter how many runs)
+#    chem_df=googleio.ChemicalData()  #Grabs relevant chemical data frame from google sheets (only once no matter how many runs)
+    chem_df = pd.read_csv('chemdf.csv')
 
-    createjson.download_experiment_directories(args.local_directory, debug)
-    target_naming_scheme = jsontocsv.printfinal(args.local_directory, debug, args.raw, chem_df)
+#    createjson.download_experiment_directories(args.local_directory, debug)
+#    target_naming_scheme = jsontocsv.printfinal(args.local_directory, debug, args.raw, chem_df)
 
-    target_naming_scheme = 'perovksitedata_20200117' # testing 
+    target_naming_scheme = 'perovskitesdata_20191209b' # testing 
     if args.verdata is not None:
         export_to_repo.prepareexport(target_naming_scheme, args.state, link, args.verdata)
 
     if args.export_reagents is True:
-        modlog.info(f'Exporting {target_naming_scheme}_models.csv and {target_naming_scheme.split}_objects.csv')
+        modlog.info(f'Exporting {target_naming_scheme}_models.csv and {target_naming_scheme}_objects.csv')
         versioned_df = export_to_repo.prepareexport(target_naming_scheme, args.state, link, args.verdata)
-        compound_ingredients.all_unique_ingredients(versioned_df, target_naming_scheme, chem_df)
+        reagent_entity.all_unique_ingredients(versioned_df, target_naming_scheme, chem_df)
             
     elif args.verdata == 0:
         modlog.info('No versioned data repository format generated')